CRUD Empleado Listo

+++ REST_API_GymTEC/Models/Employee_Extended.cs
@@ -11,7 +11,7 @@
         public string provincia { get; set; } = string.Empty;
         public string canton { get; set; } = string.Empty;
         public string distrito { get; set; } = string.Empty;
-        public float salario { get; set; } = 0.0f;
+        public double salario { get; set; } = 0.0f;
         public string correo { get; set; } = string.Empty;
         public string password { get; set; } = string.Empty;
         public string nombre_sucursal { get; set; } = string.Empty;
